cleaning(core): move pkg to node-utils
