+++ packages/core/__tests__/node-utils/framework.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test, jest, afterEach } from "@jest/globals";
 import { Frameworks, SupportedFrameworks } from "../../src/models/Framework";
-import { JsonPackage } from "../../src/filesystem/pkg";
+import { JsonPackage } from "../../src/node-utils/pkg";
 import * as FrameworkUtils from "../../src/node-utils/framework";
 import * as fs from "fs";
 import { mocked } from "ts-jest/utils";
