+++ packages/core/src/filesystem/index.ts
@@ -1,2 +1,1 @@
-export * from "./pkg";
 export * from "./PrismicSharedConfig";
