+++ packages/core/src/node-utils/index.ts
@@ -1,3 +1,4 @@
 export * as Framework from "./framework";
 export * from "./manifest";
 export * from "./paths";
+export * from "./pkg";
