+++ packages/core/src/node-utils/pkg.ts
